Fix bug
X points are first saved, sorted in distance to the lcfs (closest to furthest). The code stops when it finds the first null with opposite Z to the x point in the lcfs. Subsequently the x points are sorted in psi.

It can occur that there are x points which are closer geometrically to the lcfs but their distance in psi is greater than the last x point that we want to save (the null with opposite z)

This fix makes sure that we save only the x points that we want.