Fix find_x_point: first sort in psi than save
Some x-points are closer geometrically to the lcfs, but not closer in psi.

Now the function computes and optimizes for all x-points, then it sorts them in order of psi. Finally it saves only up to the first x-point that has opposite z coordinate to the null on the lcfs