Bug fix in find_x_point: better threshold
The threshold for eliminating the points falling inside the lcfs was not robust enough.

Now it is more robust, since the threshold is not anymore absolute. It is now defined as the DeltaPsi of the x point with the minimum distance to the separatrix. The code takes from the closest x point onward, discarding the ones that have fallen inside the lcfs.