New function find_levels_from_P
New function. it gives the levels of psi, from the discretization of a user defined trend of the heat flux at OMP. The discretization occurs by having evenly spaced levels of P to interpolate P(r) which is the integral function 2pi r q(r) dr. Integral function P(r) is always invertible, so this works for every q(r) > 0