Fix minor bug in sol
while using wall, :lfs and :lfs_far are now identified based on psi and not from the location of the intersections on the wall. Needed because, it was possible to have both strike locations of an ofl in the same semiplane, even if the ofl was in :lfs_far domain.