Overhaul of find_psi_2nd_separatrix
In the spirit of removing "a priori specified" thresholds, this new version of the functions retrieves the psi of the 2nd separatrix geometrically, using only eqt. The improvement is that now the found psi corresponds to the "first" ofl (up to precision) which starts and finishes on opposite sides to the midplane. In other words, the surface corresponding to this psi will lies outside the second X-point.