Bug fix in find_psi_last_diverted
The function assumed that the tangency point is on the opposite side of the primary x-point. This was not always the case. Now fixed by defining if a surface is diverted or not by counting the intersections with the wall. (before one looked at the sign of the Z coordinate of the intersections)