Fix bug: find_psi_last_diverted should be agnostic of number of strike points
I tried to be smart and save a minor amount of computational time when double null. This was done by a switch that counted how many strike-points were saved (4 for DN), which has been removed. The function is far more robust now that we do not make this switch and make the function agnostic of the number of strike-points that are saved. Now one can save all strike-points without impacting this function.