Fix bug in find_x_points
When double null is not perfect, the two nulls are closer in psi than the resolution of PSI_interpolant. This resulted in a bad sorting of the x-points (primary and secondary null swapped). This is solved by checking that the Z of the primary X point has the same sign of the strike points.