Add functions to average coil and magnetic data
These are based around the existing getdata functions for CER and
TS, but rather than returning all data samples they simply return
the average. These could be further generalized in the future if
there is reason to do so. They also return the time and units which
is more than required for the the present applications.